chore: reorganise v1 tests
